+++ src/main/java/yapp/buddycon/web/member/MemberRepository.java
@@ -1,4 +1,4 @@
-package yapp.buddycon.member;
+package yapp.buddycon.web.member;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
