+++ src/main/java/yapp/buddycon/web/member/auth/AuthMemberProvider.java
@@ -1,4 +1,4 @@
-package yapp.buddycon.member.auth;
+package yapp.buddycon.web.member.auth;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
