+++ src/main/java/yapp/buddycon/web/member/auth/TokenResponseDto.java
@@ -1,4 +1,4 @@
-package yapp.buddycon.member.auth;
+package yapp.buddycon.web.member.auth;
 
 public record TokenResponseDto(
   String grantType,
