+++ src/main/resources/application.yml
@@ -11,9 +11,9 @@ spring:
     password:
     driver-class-name: com.mysql.cj.jdbc.Driver
 
-  redis:
-    host: 127.0.0.1
-    port: 6379
+#  redis:
+#    host: 127.0.0.1
+#    port: 6379
 
 
 security:
